--- test/case_filter_stream_begin_end/1.btx_log.txt
@@ -1,2 +1,0 @@
-event_1: 
-event_2:
