--- test/case_filter_stream_begin_end/1.callbacks.c
@@ -1,9 +1,0 @@
-#include <metababel/metababel.h>
-
-void btx_usr_callback_event_1(void *btx_handle,void *usr_data) {
-    btx_push_message_event_1(btx_handle);
-}
-
-void btx_register_usr_callbacks(void *btx_handle) {
-  btx_register_callbacks_event_1(btx_handle,&btx_usr_callback_event_1);
-}
