--- test/cases_filter_state_machine/1.A.btx_log.txt
@@ -1,4 +1,0 @@
-event_1: { pf_1 = 0 }
-event_1: { pf_1 = 1 }
-event_1: { pf_1 = 2 }
-event_1: { pf_1 = 3 }
