--- test/cases_filter_state_machine/1.A.btx_model.yaml
@@ -1,12 +1,0 @@
----
-:stream_classes:
-- :name: scA
-  :event_classes:
-  - :name: event_1
-    :payload_field_class:
-      :type: structure
-      :members:
-      - :name: pf_1
-        :field_class:
-          :type: integer_unsigned
-          :field_value_range: 64
