--- test/cases_filter_state_machine/1.btx_model.yaml
@@ -1,20 +1,0 @@
----
-:stream_classes:
-- :name: scC
-  :event_classes:
-  - :name: event_1
-    :payload_field_class:
-      :type: structure
-      :members:
-      - :name: pf_1
-        :field_class:
-          :type: integer_unsigned
-          :field_value_range: 64
-  - :name: event_2
-    :payload_field_class:
-      :type: structure
-      :members:
-      - :name: pf_1
-        :field_class:
-          :type: integer_unsigned
-          :field_value_range: 64
