--- test/state_machine/cases_messages_count/1.btx_model.yaml
@@ -1,6 +1,0 @@
----
-:stream_classes:
-- :name: sc1
-  :event_classes:
-  - :name: event_1
-  - :name: event_2
