+++ test/state_machine/cases_no_missing_messages/1.A.btx_log.in
@@ -1,0 +1,5 @@
+event_1: 
+event_1: 
+event_1: 
+event_1: 
+event_1:
