+++ test/state_machine/cases_no_missing_messages/1.B.btx_log.in
@@ -1,0 +1,10 @@
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2: 
+event_2:
