+++ test/state_machine/cases_no_missing_messages/1.btx_model.yaml
@@ -3,4 +3,6 @@
 - :name: sc_1
   :event_classes:
   - :name: event_1
+- :name: sc_2
+  :event_classes:
   - :name: event_2
