+++ test/state_machine/cases_no_missing_messages/1.sink_callbacks.c
